fix(runtime): sign-extend boxed integer unbox on the freestanding/native path
DECODE_INT used a LOGICAL right shift, so the ENCODE_INT/DECODE_INT round trip
was not identity for negative values. ENCODE_INT(-1) is 0xFFFFFFFFFFFFFFF8 and
the old macro decoded it as (uint64_t)-1 >> 3 = 2305843009213693951 -- the exact
value the SimpleOS WM guest printed for index_of_builtin. A boxed -1 is stored
as -1 << 3, so the unbox must sign-extend, matching what cranelift already does
with sshr_imm and what the riscv64 freestanding runtime already did by shifting
a signed spl_i64.

Fixed all ten DECODE_INT definitions (x86_64, x86_32, arm64, arm32, riscv64,
plus the shared baremetal_runtime.h and the x86_64 type/crypto helper TUs) and
the two remaining logical shifts in the hosted runtime: rt_core_numeric_arg and
rt_to_string, which rendered any boxed negative integer as a 19-digit positive.

Proof: a standalone C harness over ten values shows the old macro failing 5 of
them and the new one failing none, and reproduces 2305843009213693951 exactly.
In the guest the WM gate receipt moved from
  index_of_builtin=2305843009213693951
to
  index_of_builtin=-1
with index_of_portable=3 unchanged.

Honest scope: this does NOT clear the WM guest-render-fault. The gate still
reports reason=guest-render-fault with all five markers at 0. The fault
signature did change -- serial_log_bytes 191547 -> 197001, and new negative and
near-boundary cr2 values (0xffffffffffffff83, 0x13ffffffb6) appear alongside the
previous P+0x8 / P+0x18 pattern, consistent with a separate 32/64-bit width
defect that a sign-corrected negative now surfaces rather than hides.

Two further defects found and filed, not fixed here:
- ByteSpan.starts_with is dropped from the kernel object and silently replaced
  by a nil-returning WEAK auto-stub, reached from 21 call sites in the WM CSS
  colour and gradient path.
- text.index_of returns nil in the freestanding build because rt_index_of is
  linked nowhere in the kernel.

+++ examples/09_embedded/simple_os/arch/arm32/boot/baremetal_stubs.c
@@ -95,7 +95,7 @@ RuntimeValue arm32_harden_print_canary(void)
 #define TAG_SPECIAL 0x3U
 
 #define ENCODE_INT(v)  ((RuntimeValue)(((uint32_t)(int32_t)(v) << 3) | TAG_INT))
-#define DECODE_INT(v)  ((int32_t)((uint32_t)(v) >> 3))
+#define DECODE_INT(v)  ((int32_t)(v) >> 3)
 
 #define ENCODE_PTR(p)  ((RuntimeValue)((uint32_t)(uintptr_t)(p) | TAG_HEAP))
 #define DECODE_PTR(v)  ((void*)((uint32_t)(v) & ~TAG_MASK))
