+++ examples/09_embedded/simple_os/arch/x86_64/boot/ed25519_verify_helper.c
@@ -13,7 +13,7 @@ typedef int64_t RuntimeValue;
 #define ENCODE_INT(v)  ((RuntimeValue)(((uint64_t)(int64_t)(v) << 3) | TAG_INT))
 #define ENCODE_PTR(p)  ((RuntimeValue)((uint64_t)(uintptr_t)(p) | TAG_HEAP))
 #define DECODE_PTR(v)  ((void*)((uint64_t)(v) & ~TAG_MASK))
-#define DECODE_INT(v)  ((int64_t)((uint64_t)(v) >> 3))
+#define DECODE_INT(v)  ((int64_t)(v) >> 3)
 #define IS_INT(v)      (((uint64_t)(v) & TAG_MASK) == TAG_INT)
 #define IS_HEAP(v)     (((uint64_t)(v) & TAG_MASK) == TAG_HEAP)
 #define NIL_VALUE      ((RuntimeValue)TAG_SPECIAL)
